fungsi selesai

+++ resources/views/device/view/in.blade.php
@@ -1,5 +1,7 @@
 @extends('layouts.main') @section('content')
-    <a href="/device/create" class="btn btn-success mt-4"><i class="bi bi-plus-square"></i> Tambah Perangkat</a>
+    @if (auth()->user()->role == 'team-field')
+        <a href="/device/create" class="btn btn-success mt-4"><i class="bi bi-plus-square"></i> Tambah Perangkat</a>
+    @endif
     <table class="table table-striped table-bordered mt-4">
         <thead>
             <tr class="text-center bg-success text-white">
@@ -10,7 +12,9 @@
                 <th scope="col">Model</th>
                 <th scope="col">Tipe</th>
                 <th scope="col">Deskripsi</th>
-                <th scope="col">Action</th>
+                @if (auth()->user()->role == 'field-manager')
+                    <th scope="col">Action</th>
+                @endif
             </tr>
         </thead>
         <tbody>
@@ -27,13 +31,16 @@
                     <td>{{ $d->model }}</td>
                     <td>{{ $d->type }}</td>
                     <td>{{ $d->desc }}</td>
-                    <td>
-                        <a href="#" class="btn btn-warning"><i class="bi bi-pen"></i></a>
-                        <a href="/device/delete/{{ $d->device_id }}" class="btn btn-danger"><i
-                                class="bi bi-trash"></i></a>
-                        <a href="/device/verify/{{ $d->device_id }}" class="btn btn-success"><i
-                                class="bi bi-check2-circle"></i></a>
-                    </td>
+                    @if (auth()->user()->role == 'field-manager')
+                        <td>
+                            <a href="/device/edit/{{ $d->device_id }}" class="btn btn-warning"><i
+                                    class="bi bi-pen"></i></a>
+                            <a href="/device/delete/{{ $d->device_id }}" class="btn btn-danger"><i
+                                    class="bi bi-trash"></i></a>
+                            <a href="/device/verify/{{ $d->device_id }}" class="btn btn-success"><i
+                                    class="bi bi-check2-circle"></i></a>
+                        </td>
+                    @endif
                 </tr>
             @endforeach
         </tbody>
